feat: Changing existing CPU code in PDepth2.cs to Compute shader and CG shader code to utilise GPU processing/power
If you run TEST.unity scene file, you will see a square grid of red
dots appearing in front of the Oculus Rift camera. We added a Camera
matrix taken from the Oculus Camera into the CG shader so that the
rendering of the particle grid points will follow the camera
orientation and movements.