fix: Improve the hand gesture tracking
Now the GUI-HUD can be shown/unshown by waving your hands in front
of the camera. You can do all sorts of waving, there is no fixed motion
of how the waving should start and end. After waving, if the user puts
his/her hands out, the chances of it being tracked is almost 100% as
the waving causes the camera to keep track of whatever is moving and
forces it to guess that they are hands moving.
So the waving is a gesture for users to turn on the HUD and also turn
it off, doing so also tricks the user in giving time and cues for the
system to track the hands which the camera is more than capable of
for waving hands. Next would be to implement the distance change with
the pose gestures.