feat: Add in depth distance variations with gesture poses
Now with the participant showing thumbs up, the depth distance will increase overtime; and with thumbs down, the depth distance will decrease overtime.
With open hands/palms, the depth distance will neither be increasing or
decreasing. But still the participant has to wave first to get the camera
to track the participant's hands. Of course, this trick is to get participants
thinking about displaying the GUI, and not giving time for the camera to
detect the participants' hands.