feat: Increase the depth distance increments, add in visual HUD replacing text
Gun mentioned that increasing the depth distance by 1, and decreasing by 1 is too slow and should be faster so the increment value is now 5. Also a text
HUD is very informative from a programmer's point of view but in the eyes of
participants this would be information that do not want to know if they dont need to, a picture might be a better way to visually present the information.
This makes the system application more appealing to use and fun. All HUD
functionality is done in the GestureFeedback.cs script. And we used sprites
which are generated by importing black and white images with a transparent
background.