fix: Improve on fading in/out feature of Gesture sprites
The fading in and out of the gesture pose sprites is now working!
Also the depth distance is not allowed to exceed 1550 or -50, and
the bug of not being able to increase or decrease the distance when
it exceeds the limits is now fixed. So even if the depth distance
exceeds 1550 by a bit, participants can still decrease the depth
distance.