fix: make final fixes to visual HUD
The HUD now works with waving motions, and use the thumb and open
hand gestures to increase/decrease/stop the depth distance increments.
This also works for the up and down arrow keys and spacebar to pause.
Also the HUD has been cleaned up and text all removed except for the depth
distance value which is good to keep as reference. It is now placed at
the bottom as I recall from a research paper that having something placed
at the bottom helps reduce the motion sickness/giddiness from using the
Oculus Rift head-mounted display.